Fixed a bug in convert_floating_decimal_to_hex

+++ include/core.h
@@ -4,7 +4,7 @@
 double decimal_reciprocal(int n, int pos);
 double compute_3_4_first_sum(int n, int base, int c, int (*p)(int));
 double compute_3_4_second_sum(int n, int base, int c, int (*p)(int));
-int mod_one(double x);
+double mod_one(double x);
 char* convert_floating_decimal_to_hex(double x, int places);
 
 #endif // CORE_H
